# backend/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProviderSecurityEventRequest": {
        "properties": {
          "action": {
            "enum": [
              "password_changed",
              "logout_all",
              "account_disabled"
            ],
            "title": "Action",
            "type": "string"
          },
          "eventId": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Eventid",
            "type": "string"
          },
          "issuer": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Issuer",
            "type": "string"
          },
          "occurredAt": {
            "exclusiveMinimum": 0.0,
            "title": "Occurredat",
            "type": "integer"
          },
          "subject": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "eventId",
          "occurredAt",
          "issuer",
          "subject",
          "action"
        ],
        "title": "ProviderSecurityEventRequest",
        "type": "object"
      },
...
      "SessionResponse": {
        "properties": {
          "accountId": {
            "title": "Accountid",
            "type": "string"
          },
          "issuer": {
            "title": "Issuer",
            "type": "string"
          },
          "profileId": {
            "title": "Profileid",
            "type": "string"
          },
          "sid": {
            "title": "Sid",
            "type": "string"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "accountId",
          "profileId",
          "issuer",
          "subject",
          "sid"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "KeycloakAccessToken": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "LockdIn Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/backchannel-logout": {
      "post": {
        "operationId": "backchannel_logout_api_v1_auth_backchannel_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid provider event"
          },
          "503": {
            "description": "Authentication service unavailable"
          }
        },
        "summary": "Backchannel Logout",
        "tags": [
          "authentication"
        ]
      }
    },
...
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required"
          },
          "503": {
            "description": "Authentication service unavailable"
          }
        },
        "security": [
          {
            "KeycloakAccessToken": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "operationId": "logout_all_api_v1_auth_logout_all_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required"
          },
          "503": {
            "description": "Authentication service unavailable"
          }
        },
        "security": [
          {
            "KeycloakAccessToken": []
          }
        ],
        "summary": "Logout All",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/provider-events": {
      "post": {
        "operationId": "provider_event_api_v1_auth_provider_events_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-LockdIn-Event-Signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Lockdin-Event-Signature"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderSecurityEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid provider event"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Authentication service unavailable"
          }
        },
        "summary": "Provider Event",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "operationId": "session_api_v1_auth_session_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required"
          },
          "503": {
            "description": "Authentication service unavailable"
          }
        },
        "security": [
          {
            "KeycloakAccessToken": []
          }
        ],
        "summary": "Session",
        "tags": [
          "authentication"
        ]
      }
    },
...
  }
}